// Application/Frontend/frontend-app/src/generated/trade-client/model/instrument-response.ts
/**
 * Trade REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InstrumentResponse { 
    symbol: string;
    name: string;
    price: number;
    change: number;
    changePercent: number;
}